PARQUET-1519: [C++] Hide TypedColumnReader implementation behind virtual interfaces, remove use of "extern template class"
This leaves the public API for reading columns unchanged. This avoids linking brittleness with "extern template class" and makes the ABI more stable, for whatever that's worth

If this is deemed uncontroversial (and I hope that it is) I'll do the same for the column writers in a separate patch

Author: Wes McKinney <wesm+git@apache.org>

Closes #3525 from wesm/PARQUET-1519 and squashes the following commits:

632ccded8 <Wes McKinney> Fix benchmarks, clang warnings
f078769c3 <Wes McKinney> cruft
fd99f499b <Wes McKinney> Hide TypedColumnReader implementation behind virtual interfaces, remove use of extern template class
